backend/pyproject.toml
[project]
name = "dovi-backend"
version = "0.0.1"
description = "DOVI backend — FastAPI + LlamaIndex + faster-whisper"
requires-python = ">=3.11,<3.14"
dependencies = [
    # Web
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.2.1",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    # Queue / cache
    "dramatiq[redis]>=1.17.1",
    "redis>=5.2.1",
    # Vector store
    "qdrant-client>=1.12.1",
    # RAG core (sin providers de embedding para mantener instalación liviana)
    "llama-index-core>=0.12.8",
    "llama-index-llms-anthropic>=0.6.3",
    "llama-index-llms-ollama>=0.5.0",
    "llama-index-vector-stores-qdrant>=0.4.2",
    # LLM clients
    "anthropic>=0.42.0",
    # Retry / resilience
    "tenacity>=9.0.0",
    # Logging
    "structlog>=24.4.0",
    # Parsing
    "webvtt-py>=0.5.1",
    # HTTP utilities
    "httpx>=0.28.1",
]

[project.optional-dependencies]
# Embeddings + torch (pesado). Instalar con: uv sync --extra embeddings
embeddings = [
    "llama-index-embeddings-huggingface>=0.5.0",
    "FlagEmbedding>=1.3.2",
]
# ASR (faster-whisper + ctranslate2). Instalar con: uv sync --extra asr
asr = [
    "faster-whisper>=1.1.0",
]
# Diarización opcional (plan §4.17)
diarization = [
    "pyannote-audio>=3.3.2",
]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[build-system]
requires = ["setuptools>=76", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.uv]
# uv gestiona el interpreter aislado; respeta .python-version para pinning.
python-preference = "managed"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ASYNC", "S", "PL", "RUF"]
ignore = ["S101", "PLR0913"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
